+++ src/fugle.hpp
@@ -7,3 +7,5 @@
 
 #include "fugle_client_base.hpp"
 #include "fugle_intraday.hpp"
+
+#include "fugle_api_list.hpp"
