+++ src/fugle_api_list.hpp
@@ -1,6 +1,7 @@
 
 #pragma once
 
+#include <nlohmann/json.hpp>
 #include <string>
 #include <vector>
