+++ src/fugle_intraday.cpp
@@ -1,11 +1,13 @@
 
-#include "fugle_intraday.hpp"
-#include "fugle_client_base.hpp"
-#include "fugle_intraday_data.hpp"
 #include "str_utils.hpp"
 #include <nlohmann/json.hpp>
 #include <spdlog/spdlog.h>
 #include <string>
+
+#include "fugle_client_base.hpp"
+#include "fugle_common_data.hpp"
+#include "fugle_intraday.hpp"
+#include "fugle_intraday_data.hpp"
 
 using namespace std;
 using namespace fugle_realtime;
