+++ src/str_utils.hpp
@@ -1,7 +1,9 @@
 
 #pragma once
 #include <map>
+#include <sstream>
 #include <string>
+#include <vector>
 
 namespace fugle_realtime {
